+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/statistic/ArrowheadFP.java
@@ -29,7 +29,7 @@ public class ArrowheadFP implements Statistic {
     @Override
     public double getValue(Graph trueGraph, Graph estGraph, DataModel dataModel) {
         ArrowConfusion confusion = new ArrowConfusion(trueGraph, estGraph);
-        return (double) confusion.getArrowsFp();
+        return confusion.getArrowsFp();
     }
 
     @Override
