+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/statistic/ArrowheadTN.java
@@ -29,7 +29,7 @@ public class ArrowheadTN implements Statistic {
     @Override
     public double getValue(Graph trueGraph, Graph estGraph, DataModel dataModel) {
         ArrowConfusion confusion = new ArrowConfusion(trueGraph, estGraph);
-        return (double) confusion.getArrowsTn();
+        return confusion.getArrowsTn();
     }
 
     @Override
